add tests to orders table

+++ migrations/sqls/20211028224423-order-products-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE order_products(
   id SERIAL PRIMARY KEY,
   quantity INTEGER,
-  order_id BIGINT REFERENCES orders(id) NOT NULL,
-  product_id BIGINT REFERENCES products(id) NOT NULL
+  order_id INTEGER REFERENCES orders(id) NOT NULL,
+  product_id INTEGER REFERENCES products(id) NOT NULL
 );
